Fix CI: modernize workflow, fix clippy/sort violations, repair no_std build
- Replace archived actions-rs/clippy-check and styfle/cancel-workflow-action
  (deprecated Node runtimes) with plain cargo clippy and a concurrency group
- Install cargo-sort via taiki-e/install-action (prebuilt binary) instead of
  compiling from source, which was failing the job in under a minute
- Run cargo sort with --workspace so member manifests are actually checked,
  and sort them
- Lint tests too (--all-targets) and fix the two unnecessary u64 casts
- Test with --all-features so the serde path is exercised
- Fix no_std build: f64::ln does not exist in core; compute the unit
  exponent with an exact integer loop instead. Drop the duplicate no_std
  attribute. Add a no-default-features check step to CI.

+++ bytescale/src/lib.rs
@@ -44,7 +44,7 @@ mod tests {
     fn test_arithmetic_primitives() {
         let mut x = ByteScale::mb(1);
 
-        assert_eq!((x + MB as u64).as_u64(), 2_000_000);
+        assert_eq!((x + MB).as_u64(), 2_000_000);
 
         assert_eq!((x + MB as u32).as_u64(), 2_000_000);
 
@@ -60,7 +60,7 @@ mod tests {
 
         assert_eq!((x - B as u32).as_u64(), 999_999);
 
-        x += MB as u64;
+        x += MB;
         x += MB as u32;
         x += 10u16;
         x += 1u8;
